+++ src/benchmark/java/com/commercetools/sync/benchmark/CategorySyncBenchmark.java
@@ -1,6 +1,5 @@
 package com.commercetools.sync.benchmark;
 
-import com.commercetools.sync.commons.utils.SyncSolutionInfo;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -17,19 +16,19 @@ public class CategorySyncBenchmark {
     @Test
     public void sync_NewCategories_ShouldCreateCategories() throws IOException {
         // TODO: SHOULD BE IMPLEMENTED.
-        saveNewResult(SyncSolutionInfo.LIB_VERSION, CATEGORY_SYNC, CREATES_ONLY, 20000);
+        saveNewResult(CATEGORY_SYNC, CREATES_ONLY, 20000);
     }
 
     @Test
     public void sync_ExistingCategories_ShouldUpdateCategories() throws IOException {
         // TODO: SHOULD BE IMPLEMENTED.
-        saveNewResult(SyncSolutionInfo.LIB_VERSION, CATEGORY_SYNC, UPDATES_ONLY, 10000);
+        saveNewResult(CATEGORY_SYNC, UPDATES_ONLY, 10000);
     }
 
     @Test
     public void sync_WithSomeExistingCategories_ShouldSyncCategories() throws IOException {
         // TODO: SHOULD BE IMPLEMENTED.
-        saveNewResult(SyncSolutionInfo.LIB_VERSION, CATEGORY_SYNC, CREATES_AND_UPDATES, 30000);
+        saveNewResult(CATEGORY_SYNC, CREATES_AND_UPDATES, 30000);
     }
 
 }